Update InterestPoint model + add factory for points requests + add async/await mechanism and error handling
